Minor reorganization, renaming.
Renamed exception and stream class to JSONError and JSONStream respectively.
JSONStream class no longer inherits directly from StringIO.  This is to ensure
that only methods on JSONStream are used by the parser, or any other parser
which extends this.  Converted global function skipspaces() into a method on
the stream.  Generalized skipping forward in the stream.