Object string conversion and encode/decode aliases.
Now falls back to casting objects which implement '__str__', to support
JSON-encoding objects which do not inherit from base types but which have
string representations.  Unfortunately JSON is not a true object serialization
format so these objects cannot be decoded to their original classes without
special handling which is not covered in the JSON spec.

Also added aliases "encode = to_json" and "decode = from_json" to match
the signatures of other JSON modules.